Implement new VencordNative for DiscordWeb
